Feat: admin인증 추가

+++ src/main/java/com/udoolleh/backend/web/RestaurantController.java
@@ -1,12 +1,17 @@
 package com.udoolleh.backend.web;
 
 import com.amazonaws.Response;
+import com.udoolleh.backend.exception.errors.CustomJwtRuntimeException;
+import com.udoolleh.backend.provider.security.JwtAuthTokenProvider;
+import com.udoolleh.backend.provider.service.AdminAuthenticationService;
 import com.udoolleh.backend.provider.service.MenuService;
 import com.udoolleh.backend.web.dto.CommonResponse;
 import com.udoolleh.backend.web.dto.RequestMenu;
 import com.udoolleh.backend.web.dto.ResponseMenu;
 import com.udoolleh.backend.web.dto.*;
 
+import java.util.Optional;
+import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import com.udoolleh.backend.core.type.PlaceType;
 import com.udoolleh.backend.core.type.UdoCoordinateType;
@@ -44,10 +49,16 @@ public class RestaurantController {
     private final MenuService menuService;
     private final KakaoApiService kakaoApiService;
     private final RestaurantService restaurantService;
+    private final JwtAuthTokenProvider jwtAuthTokenProvider;
+    private final AdminAuthenticationService adminAuthenticationService;
 
-    @PostMapping("/admin/restaurant/menu")
-    public ResponseEntity<CommonResponse> registerMenu(@RequestPart MultipartFile file,
+    @PostMapping("/restaurant/menu")
+    public ResponseEntity<CommonResponse> registerMenu(HttpServletRequest request, @RequestPart MultipartFile file,
                                                        @Valid @RequestPart RequestMenu.RegisterMenuDto requestDto){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
         menuService.registerMenu(file, requestDto);
 
         return ResponseEntity.ok().body(CommonResponse.builder()
@@ -55,7 +66,7 @@ public class RestaurantController {
                 .build());
     }
     @GetMapping("/restaurant/{id}/menu")
-    public ResponseEntity<CommonResponse> getMenu(@PathVariable String id){
+    public ResponseEntity<CommonResponse> getMenu(HttpServletRequest request, @PathVariable String id){
         List<ResponseMenu.MenuDto> list = menuService.getMenu(id);
 
         return ResponseEntity.ok().body(CommonResponse.builder()
@@ -64,8 +75,12 @@ public class RestaurantController {
                 .build());
     }
 
-    @DeleteMapping("/admin/restaurant/{id}/menu/{name}")
-    public ResponseEntity<CommonResponse> deleteMenu(@PathVariable String id, @PathVariable String name){
+    @DeleteMapping("/restaurant/{id}/menu/{name}")
+    public ResponseEntity<CommonResponse> deleteMenu(HttpServletRequest request, @PathVariable String id, @PathVariable String name){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
 
         menuService.deleteMenu(id, name);
 
@@ -74,8 +89,13 @@ public class RestaurantController {
                 .build());
     }
 
-    @PostMapping("/admin/kakao/restaurant")
-    public ResponseEntity<CommonResponse> registerRestaurantInfo(@RequestBody Map<String, PlaceType> placeType){
+    @PostMapping("/kakao/restaurant")
+    public ResponseEntity<CommonResponse> registerRestaurantInfo(HttpServletRequest request, @RequestBody Map<String, PlaceType> placeType){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
+
         kakaoApiService.callKakaoApi("우도",1,placeType.get("placeType"),UdoCoordinateType.ONE_QUADRANT); //1사분면 저장
 
         return ResponseEntity.ok().body(CommonResponse.builder()
@@ -83,9 +103,14 @@ public class RestaurantController {
                 .build());
     }
 
-    @PostMapping("/admin/restaurant/images")
-    public ResponseEntity<CommonResponse> registerRestaurantImage(@RequestPart(value="images") List<MultipartFile> images,
+    @PostMapping("/restaurant/images")
+    public ResponseEntity<CommonResponse> registerRestaurantImage(HttpServletRequest request, @RequestPart(value="images") List<MultipartFile> images,
                                                                   @RequestPart(value="restaurantName") Map<String, String> restaurantName){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
+
         restaurantService.registerRestaurantImage(images, restaurantName.get("restaurantName"));
 
         return ResponseEntity.ok().body(CommonResponse.builder()
@@ -93,8 +118,13 @@ public class RestaurantController {
                 .build());
     }
 
-    @PostMapping("/admin/restaurant")
-    public ResponseEntity<CommonResponse> registerRestaurant(@Valid @RequestBody RequestRestaurant.RegisterRestaurantDto registerDto){
+    @PostMapping("/restaurant")
+    public ResponseEntity<CommonResponse> registerRestaurant(HttpServletRequest request, @Valid @RequestBody RequestRestaurant.RegisterRestaurantDto registerDto){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
+
         restaurantService.registerRestaurant(registerDto);
 
         return ResponseEntity.ok().body(CommonResponse.builder()
@@ -113,8 +143,13 @@ public class RestaurantController {
                 .build());
     }
 
-    @DeleteMapping("/admin/restaurant/{id}/images")
-    public ResponseEntity<CommonResponse> deleteRestaurantImages(@PathVariable String id){
+    @DeleteMapping("/restaurant/{id}/images")
+    public ResponseEntity<CommonResponse> deleteRestaurantImages(HttpServletRequest request, @PathVariable String id){
+        Optional<String> token = jwtAuthTokenProvider.resolveToken(request);
+        if(!adminAuthenticationService.validAdminToken(token.orElseThrow(() -> new CustomJwtRuntimeException()))) {
+            throw new CustomJwtRuntimeException();
+        }
+
         restaurantService.deleteRestaurantImage(id);
 
         return ResponseEntity.ok().body(CommonResponse.builder()
